Add class for calculating informed moments
Add a class which functions as an objective functional model of
biased weighted moments and its derivatives given input moments
and weights. This is intended to be used in conjunction with an
optimizer to find the moments which maximize the likelihood of
moments given input measurements from data. Add files to repository
necessary for building the code.